Add two files that need to be installed
The vi script and the vimrc need to be copied to /usr/bin/ and /etc/,
respectively.

We will need to bundle them as a proper MSys2 package so that Git for
Windows can depend on that package. That way, it will be very easy
to use pacman to figure out the list of files to include in the
installer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>