Start writing a helper to compare thickets
A "thicket" is a branch of branches, i.e. a branch that contains
several topics with accompanying merge commits.

In Git for Windows, we use such a thicket to organize the
Windows-specific patch series in such a way that we can easily submit
them upstream when they are ready.

Everytime Git for Windows synchronizes with a new upstream version,
a so-called "merging rebase" is performed, keeping the thicket structure
intact. Typically, minor changes have to be made when rebasing, e.g.
when a test case is added in the same place an upstream patch added
a test case.

The big problem is that it is relatively hard to see what changed
between such merging rebases.

The compare-thickets script being started in this commit will hopefully
help with such a comparison, by giving an overview of the real changes
(i.e. listing only commits that are really different between thicket
revisions).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>