Changed pacman install command.
Due to the requirement to run post install scripts, the pacman -Syu
command to install the build dependencies cannot be run in the windows
cmd window. The initial msys2-runtime update runs fine because the lack
of post-install scripts.

After trying a solution with a second shell script that runs the install
we settled on this solution by the sh.exe -c ability.

Signed-off-by: nalla <nalla@users.noreply.github.com>