pacman-mirror: Support MinGW packages, too
As pointed out by Sebastian Schuberth, we should make a difference
between MSys2 and MinGW packages.

Happily, we can deduce from the "mingw-w64-" prefix of package names
whether we are looking at a MinGW package or not.

Therefore, let's keep MSys2 and MinGW packages separate, even if it
comes at the price of substantial additional nesting.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>