pacman-mirror: generate only two repositories, not four
Earlier, we tried to sort packages into a very fine-grained hierarchy,
but every repository needs to be added to pacman.conf separately,
and its package index needs to be synchronized separately, too,
therefore that fine-grained hierarchy not only complicated the
pacman-mirror script but also caused a lot of unnecessary churn.

Sadly, we cannot have a single flat repository, though, because the
package index cannot hold packages for different architectures under the
same name. We are thus still forced to have a separate repository for
both architectures we support: i686 and x86_64.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>