Add the nsswitch.conf file, needed to set $HOME correctly
The Cygwin way -- and thus the MSys2 way -- to define how the getpw*()
family of functions are accomodated is to interpret the nsswitch.conf
file. Since we really want to use the users' home directories rather
than /home/<user>/, we have to tell MSys2 to use Windows to fill the
entries of the 'passwd' struct.

Pointed out by @Alexpux, the MSys2 maintainer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>