git-extra: Relocation of the git-extra package
The splitup of the package file and the source was not such a good idea.
It was decided that the package file should be placed with the source
files. The `PKGBUILD` was adapted to handle the source files relative to
its location. The old location of this packe was in the
https://github.com/git-for-windows/msys2-packages repository.

Signed-off-by: nalla <nalla@hamal.uberspace.de>