Add the `git-extra.install` file
Technically, we should not track this file because it is generated.
However, `makepkg-mingw` refuses to run any function in `PKGBUILD`
that could generate it if the file is not there yet. Catch-22.

So let's just give in and track that file (and all future revisions).
It is not *that* bad, as we track auto-updates to the `pkgver`
variable in the `PKGBUILD` file already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>