Add the optional dependency on 'filesystem'
According to PKGBUILD(5), the optdepends directive is "currently for
informational purposes only" and is "not utilized by pacman during
dependency resolution". However, just in case that pacman will use it to
determine the order in which packages should be installed one day, we
record the `filesystem` package as dependency to make sure that we can
override `nsswitch.conf` and friends.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>