Override MSys2's PS1 already in git-extra's post-install script
MSys2 unfortunately hardcodes a Bash prompt that we want to override. We
therefore have to edit the /etc/bash.bashrc file. So far, we did that in
the /etc/profile.d/git.sh script, but that is overkill: it needs to be
done exactly once. Therefore, the post-install/post-upgrade code is
exactly the right place to do it -- if needed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>