Edit pacman.conf in git-extra's post-install script, too
We need to override a couple of MSys2 packages, at least for the moment.
Therefore we need to force-overwrite the `pacman.conf` configuration
from the `pacman` package.

The best place for that override seems to be the post-install script of
the git-extra package. Technically, `pacman` should be added to the
`optdepends` directive, but you cannot install the `git-extra` package
if `pacman` has not been installed first... ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>