Copy-edit msysGit's net installer setup script
It looks quite different now, actually more like a rewrite, because

- we use Pacman instead of Git to keep our development environment
  up-to-date these days.

- we dropped Git Cheetah (too many problems, too few developers).

- we want to generate our HTML documentation ourselves, now that we have
  a working asciidoc.

Unfortunately, this commit also drops support for HTTP proxies for now:
It is unclear how Pacman deals with proxies so far. Most likely, it
heeds the `http_proxy` variable just the same as Git does (because both
rely on cURL to do the actual job); This needs work by a contributor
who is behind an HTTP proxy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>