git-extra: make the .install file independent of the build architecture
We will need to commit the generated git-extra.install file soon because
makepkg refuses to start without that file. That means that we need to
avoid having two different versions for the different architectures we
support.

In any case, it makes sense to write out the gitconfig file as soon as
we know that the respective etc/ directory exists.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>