WIP: try to work around clone failures
For some reason, this developer experiences a lot of clone failures. The
tell-tale is that the exit code is 128, but sometimes, confusingly, the
clone works, and the exit code is 0.

It appears that trying to debug this by setting GIT_CURL_VERBOSE=1
(possibly in conjunction with GIT_TRANSPORT_HELPER_DEBUG=1) makes it
more likely to pass.

It could well be that this bug is related to another bug: The progress
while the clone transfers objects is not shown (probably because of a
failing isatty() call).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>