net-installer: Make "-c" an option to clone, not to git itself
If passed to git, the option is documented to "pass a configuration
parameter to the command". However, as clone is creating the initial
version of .git/config, it is not reading any configuration parameter
(neither from .git/config nor the command line), but writing them. Thus,
the correct way is to use "-c" with the clone command, which is documented
to "set a configuration variable in the newly-created repository".

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>