pacman-mirror add: remove obsoleted versions of the added files
When `repo-add` runs, it will try to add obsolete file versions first,
if present, and later replace them with the current version. That is a
lot of unneeded churn, so let's just remove previous versions when
adding new ones.

Note: this only affects the local mirror; Bintray will still have the
old versions, so we naturally have a backup.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>