net installer: ask for an HTTP proxy upon pacman failure
Ideally, we would test for a good indicator that a proxy is required
before we ask for an HTTP proxy. Alas, this developer has no access to a
setup that requires a proxy to access the web, and was unable to get
ahold of that information otherwise.

So let's just simply always ask for an HTTP proxy when pacman failed to
synchronize and/or update the msys2-runtime package.

For good measure, add a script to etc/profile.d/ that will configure the
proxy for Git Bash sessions, but only when we set the environment
variables ourselves; If the http_proxy variable was set appropriately
already outside of the installation script, we do not need to tell the
Bash to set them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>