portable/release: automaticaly insert the bitness
Most Windows computers seem to be 64-bit these days, and with version 2.x,
Git for Windows users on those computers do not need to use a 32-bit
version anymore.

To make it more obvious to users which file to download and install, the
file name should already state whether it offers the 32-bit or the 64-bit
version of Git for Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>