git-extra: Let's not always cd to the home directory
MSys2's `filesystem` package installs a post-install script that tries
to make the home directory and switch to it upon startup.

In Git for Windows' context, we want to avoid both: we want to use what
Windows thinks is the user's home directory (and that directory exists
long before the user can use Git for Windows). We also want to give
third-party programs such as TortoiseGit or GitHub for Windows a chance
to open a Git Bash in a given directoy, so we should not force the
directory to be $HOME upon `bash --login -i` (or `git-bash.exe`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>