Copy msysGit's script to generate an end-user installer
In the next few commits, we will adapt this to the MSys2-based Git for
Windows, preparing for the first release of Git for Windows 2.x.

This commit copies the entire /share/WinGit/ directory wholesale from
msysGit as of commit 4b4796f5(installer: Fix an endless loop for paths
with only spaces (issue 335)).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>