installer: port to *MSYS2*
This is a rough port of the inno setup files to the current *MSYS2*
environment. The source files are generated by the `release.sh` into
`file-list.iss`. That file is then included by adding the line
`#include "file-list.iss"` into `install.iss.in`.

The git-cheeta plugin is commented out temporary.

Signed-off-by: nalla <nalla@hamal.uberspace.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>