installer: Use the `git-bash.exe` executable
We no longer require a VisualBasic script for essentially the same task
for which we introduced the Git wrapper. We use a slightly modified copy
of the Git wrapper now, for maximum flexibility. For example, the
`git-bash.exe` now correctly identifies the Git version, and we will
most likely start signing the executables at some stage, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>