installer: generate bindimage.txt
The Win32 API provides a very useful call to optimize loading
executables and shared libraries, called BindImage(). Already in Git for
Windows 1.x' time, we called that function to make startup of Git a
little bit swifter.

In Git for Windows 2.x' time, we need to ask Pacman for the executables
and shared libraries, though, not `find`, because we do not populate a
fresh root directory any longer but instead pick the files from a
running Git SDK that need to be included into the installer -- a `find
-name *.exe` would find also files that will not be installed onto end
users' machines.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>