installer: fall back to copying the Git wrapper when hard-linking fails
The idea of Git's builtins is to provide backwards compatibility by hard
linking the `git.exe` binary to `git-<name>.exe` for all of the Git
commands that were converted to builtins.

However, when hard-linking fails, copying the `git.exe` over a hundred
times is quite costly: every copy takes 7-8 megabytes at time of writing,
tendency going up.

For this reason, the Git for Windows project extended the Windows-specific
Git wrapper (that essentially executes the *real* `git.exe` after setting
up some environment variables, i.e. does something similar to what we
need) to serve as a drop-in replacement for all of those builtins.

Let's use it if hard-linking fails.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>