pacman-mirror: use the certificates of the 'ca-certificates' bundle
It appears to this developer as if the certificates shipped with the
mingw-w64-*-ca-certificates packages lag behind and ca-certificates is
much more up-to-date. At least twice were certificate issues with
Bintray solved by using the certificates bundle from the latter package,
despite being fully up-to-date on both ca-certificates packages.

So let's just hard-code the path to the bundle that is most likely to be
up-to-date.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>