installer: syntax fixes and performance improvements
The generation of files to be shipped by the installer was significantly
improved by using a `sed` call. Also the `release.sh` is a little more
chatty now. The git tag is now only done if the tag does not exists
already. There were some syntax errors in the `installer.iss.in` file
that were squashed.

Helped-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: nalla <nalla@hamal.uberspace.de>