Add a PKGBUILD for Markdown
Previously, Git for Windows' release notes were provided as an .rtf
file. The intention was to enable somewhat more pleasant formatting than
plain text.

However, the .rtf format is not without flaws: re-saving the same file
in a different editor will change the file on disk even if no textual
content nor formatting was changed. Besides, the formatting of an .rtf
file is sadly not as standardized as we hoped: in Windows 8's WordPad,
the bullet points are displayed differently than in XP's WordPad.

The markdown standard provides all the formatting we would want to have,
and it can be converted easily into HTML which can be displayed just
after Git for Windows was installed. So let's package the Markdown
package for use in any MSys2-based system -- including our Git SDK --
and use it from now on to format our release notes.

To accomodate Git for Windows' Pacman repository structure, we build
architecture dependent versions of the package, even if Markdown is
technically not architecture dependent (by virtue of being a Perl
script).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>