installer: convert the release notes to HTML
The `.rtf` format we used to present our release notes is sadly not as
much of a standard as we hoped, as it is rendered differently depending
whether Word or WordPad, or even which version thereof, is used to
display it.

As we already make extensive use of markdown on GitHub (README, wiki,
comments, etc), it makes sense to maintain the release notes in markdown
format, too.

While at it, add a few changes since 1.9.5 already, even if the next Git
for Windows version needs to wait a few days, still.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>