installer: Add installer page for performance tweaks
The core.fscache config option provides a significant performance boost,
but it's still experimental. Since wider use can help flush out any
issues and push this from being experimental to production-ready, this
adds a page to the installer advising users of the setting's existence
and allowing them to enable it.

Signed-off-by: Eli Young <elyscape@gmail.com>