Make make-file-list.sh useful, then use it in installer/release.sh
The idea of make-file-list.sh was originally to generate a list given a
list of packages. Of course its intended use was in the script
generating the installer or the portable application.

It is time to reinstate that original purpose. While at it, hard-code
the package list in `make-file-list.sh`, accepting optional packages as
command-line parameters.

Incidentally, this consolidates the corresponding code in the scripts
generating the portable application and the installer, respectively.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>