make-file-list: make the list more similar to msysGit's
Whether using the portable application or using the installer to set up
a working Git for Windows (for end users), it is quite likely that the
SSH transport needs to be used.

After this had been pointed out in a very, very brief issue report that
ssh seems to be missing from the portable application, a thorough
investigation was performed to figure out the differences between the
1.9.5 portable application and the new one: The output of

7za l PortableGit-1.9.5-preview20141217.7z.exe |
sed -n 's/^.\{20\}\..\{32\}//p' |
sed -e 's|^mingw|mingw64|' \
    -e 's|^bin|usr/bin|' \
    -e 's|^doc/git/html|mingw64/share/doc/git-doc|' \
    -e 's|^etc/git[a-z]|mingw64/&|' \
    -e 's|^lib/engines/|usr/lib/openssl/engines/|' \
    -e 's|^lib/perl5/5.8.8/|usr/share/perl5/core_perl/|' \
    -e 's|^lib/perl5/site_perl/5.8.8/msys/|usr/lib/perl5/vendor_perl/|' \
    -e 's|^lib/perl5/site_perl/|mingw64/share/perl5/site_perl/|' \
    -e 's|^lib/tcl8.5/|mingw64/lib/tcl8.6/|' \
    -e 's|^lib/tk8.5/|mingw64/lib/tk8.6/|' \
    -e 's|^lib/tcl|mingw64/&|' \
    -e 's|^lib/tk|mingw64/&|' \
    -e 's|^lib/dde1.3|mingw64/lib/dde1.4|' \
    -e 's|^libexec/git-core/|mingw64/libexec/git-core/|' |
    sort |
    uniq

and of

7za l PortableGit-2.3.5.3-preview.7z.exe |
sed -n 's/^.\{20\}\..\{32\}//p'

were compared. As a consequence, (parts of) a couple of packages were
excluded explicitly.

Oh, and ssh.exe is now included.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>