installer: include HTML pages again
It has been (kind of) pointed out in a Pull Request that the exclusions
when generating a list of files to include into the installer went too
far, excluding even the Git help pages.

This was of course unintended.

To fix this, and to avoid further confusion let's be more explicit about
*what* share/doc/ subdirectories are to be excluded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>