git-extra: prevent MSys2 from overriding the TMP variable
In Git for Windows, we actually did not experience any trouble when
running with the user's temporary directory as TMP variable at all.

Worse, the end user will have Git for Windows installed in their Program
Files directory which means that the tmp/ variable would be in a write
protected area, and when Git wants to write into that directory when
non-administrators run a command, it would actually fail.

So let's just disable that.

Pointed out by Karsten Blees during the Git Merge 2015.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>