net-installer: Use the simpler "ExecuteFile" instead of "RunProgram"
The 7-Zip docs are not 100% clear on this. However, from reading the
docs it appears this was merely about calling CreateProcess (in case of
RunProgram) vs. ShellExecute (in case of ExecuteFile), and a look at the
SFX source code confirms this. The code also shows that ExtractArchive
is called in the same way before any of RunProgram or ExecuteFile is
evaluated.

But lastly this developer had also tried it and it worked :-)

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>