pacman-mirror: expect all MinGW packages architecture to be 'any'
Pacman is nice enough to help users by preventing them from installing
packages for an incompatible architecture. For example, on ARM Linux, you
cannot install a MIPS openssh package. The same is true for MSys2
packages: an OpenSSH for, say, i686 cannot work in an environment where an
x86_64 msys2-runtime is installed.

However, it is not helpful to prevent MinGW packages for i686 from being
installed in an x86_64 setup: these are standalone programs and they work
just fine.

That is the reason why MSys2 encodes the architecture of MinGW packages in
their package name, but marks the package's architecture as 'any' (even if
it is technically incorrect and might very well lead to problems with
Windows 10 that does run on ARM processors, where mixing and matching
different architectures is no longer a safe thing: you will not be able to
run an i686 program on an ARM processor).

For example, when building Git for Windows packages, this developer uses a
setup where *both* mingw-w64-i686-toolchain and mingw-w64-x86_64-toolchain
are installed. This works quite well and lets `makepkg-mingw` build the
package for *both* architectures in one go.

Up to now, the Git for Windows project handled the architecture of the
mingw-w64-git package incorrectly by marking the packages' architecture as
either i686 or x86_64 and thereby preventing users from installing a
mingw-w64-git package for a different architecture than the current
msys2-runtime one.

However, with the mingw-w64-asciidoctor-extensions package being a build
dependency, the problem cropped up that marking its architecture would
prevent it to be installed for both architectures, breaking the paradigm
where mingw-w64-git could be built for both architectures in one go.

Therefore, we now build our MinGW packages also with the "any"
architecture, just like upstream, and everybody is happy.

This requires the change contained in this commit, though, because the
packages are no longer named mingw-w64-i686-*-i686.pkg.tar.xz and
mingw-w64-x86_64-*-x86_64.pkg.tar.xz, respectively, but instead
mingw-w64-i686-*-any.pkg.tar.xz and mingw-w64-x86_64-*-any.pkg.tar.xz.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>