installer: honor `%PROGRAMDATA%\Git\config`
Before trying to set the selected config values in the installer the code
now checks if the system wide git config exists. If that is not the case
the installer copies the shipped config to `%PROGRAMDATA%\Git\config` and
then tries to alter it as choosen by the user.

The shipped gitconfig is deleted after checking if the system wide git
config already exists.

Github-issue: git-for-windows/git#121
Signed-off-by: 마누엘 <nalla@hamal.uberspace.de>