shears: stop trusting 'echo'
With the recent addition of the commit whose oneline reads

	Windows: use %PROGRAMDATA%\Git\config as Windows-wide configuration

dash's builtin echo(1) command started acting up, cutting off the todo
list stored in the $origtodo variable just after "\Git". This cause
substantial problems because it dropped commits inadvertently. While that
problem was detected before pushing a rebased Git for Windows, it needed
to be fixed on the script level to prevent likewise problems in the
future.

The true fix lies a little in the future: a rewrite of the shears as a
builtin (which should be based on a rewrite of the interactive rebase as a
builtin).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>