shears: add special-case handling for the 'merging-rebase' argument
This maintainer's workflow when performing a merging rebase typically
starts with an initial merging rebase:

$ shears --merging --onto=v2.3.7 <previous-merging-rebase-start>

After that, we need to verify that no inadvertent changes were introduced.
The easiest way to do that is to compare the previous diff and the current
diff vs upstream:

$ git diff v2.3.6..origin/master >prev.txt
$ git diff v2.3.7..master >cur.txt
$ git diff --no-index prev.txt cur.txt

The only changes to be expected are the ones in `index` and in `@@` lines.
If any line is changed that starts with a `-` or a `+`, something cooky is
going on and the merging rebase will most likely need to be redone

After that initial merging rebase, but still *before* pushing, this
maintainer takes the opportunity to clean up a little, i.e. reordering
commits and/or branches, merging branches into other branches, sometimes
dropping patches which we reverted before submitting upstream, etc:

$ shears <current-merging-rebase-start>
$ <lather, rinse and repeat>

To verify that this cleanup process did not drop or introduce changes, it
is enough to ensure that `git diff master@{1}` comes up empty-handed.

In both `shears` calls, it is necessary to provide the latest commit that
started a merging rebase. So let's just make that less tedious by
special-casing the value `merging-rebase` which automatically looks for
the latest such commit and substitutes the SHA-1 of that commit, then
continues as before. That way, the former call becomes

$ shears --merging --onto=v2.3.7 merging-rebase

and the latter

$ shears merging-rebase

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>