installer: Fix .sh file association
`%1` expands to the short name of the file, rather than the directory
the file is contained in, so `--cd=%1` fails when `git-bash.exe` tries
to change the working directory to a file. `%w` expands to the file's
containing directory and `%L` expands to the long name of the file.

This commit also forces update of the command string.

Helped-by: @Animeye
Signed-off-by: Eli Young <elyscape@gmail.com>