Describe cd and no-cd flags for git-bash in the release notes
The Start Menu shortcut for Git Bash supported "Start In" until 1.9.5.

With the new git-bash executable, Start In has no effect. This is done in
order to make portable git always start in the home directory rather than
the directory it is located in.

As a replacement, --cd=<dir> argument is supported by git-bash.

Describe it in the release notes for users that modified the working
directory for the shortcut.

Signed-off-by: Orgad Shaneh <orgad.shaneh@audiocodes.com>