Fix `.sh` file association on Windows XP
At least on Windows XP, the `%w` keyword is not expanded to the
directory of the file that is to be handled by the file association. As
a consequence, the `.sh` file association did not work at all, failing
silently.

But we actually do not require the directory name at all: we just need
to prevent `git-bash.exe` from switching the directory to begin with.
Let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>