installer: remove files from a previous installation, if any
In particular when upgrading from a 32-bit to a 64-bit version of Git
for Windows, we really need to remove the files from the previous
version first (because there will otherwise be left-over files in
`C:\Program Files (x86)\Git` that will never get deleted).

The "system-wide" Git config (previously in `/etc/gitconfig`) needs
special care: as agreed with the libgit2 project, we share the
Windows-wide Git configuration in `%PROGRAMDATA%\Git\config` now.

This fixes https://github.com/git-for-windows/git/issues/107.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>