installer/portable: include the `start` script
It was pointed out by Orgad Shaneh that the `start` script is not part
of the end users' Git for Windows installation.

This script is part of the `filesystem` package. As that package
contains files we do not want to ship with Git for Windows, we pluck
just the `start` script out of it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>