installer: add a page to select the terminal emulator for Git Bash
While this developer is really, really happy with MinTTY, not everybody
agrees. In the fine tradition of adding new wizard pages to the
installer to let everybody configure Git for Windows the way they like
it best, this commit adds a page that lets the person installing Git for
Windows decide what terminal emulator is best, defaulting to MinTTY, and
offering to opt for Windows' default console window instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>