installer: avoid extra console window with Git Bash
As reported by Valerii Pekarskyi, when selecting the ConHost-based Git
Bash, we had an extra console window. The reason is that we added the
SHOW_CONSOLE=1 configuration to `git-bash.exe` *in addition to* making
`git-bash.exe` a console application (by copying over the `git-cmd.exe`,
overwriting `git-bash.exe`). Let's do only the former and skip the
latter.

This fixes https://github.com/git-for-windows/git/issues/148.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>