installer: Avoid git for windows DLLs conflicts with system ones
When an executable requires a `.dll` to be loaded, Windows first looks
in the same directory as the executable, then in the system directory
("C:\Windows\system32\"), and only long after that in the `PATH`. See
https://msdn.microsoft.com/en-us/library/windows/desktop/ms682586 for
details.

Now, when e.g. `git-remote-https.exe` is the executable in question and
the corresponding `.dll` is `ssleay32.dll`, there is a problem if that
`.dll` file exists in the system directory: `git-remote-https.exe` lives
in `/mingw64/libexec/git-core/` (or `/mingw32/...` in 32-bit setups),
while the `ssleay32.dll` library is in `/mingw64/bin/`, and therefore
Windows will find the wrong one.

That is particularly problematic if the version of the library does not
match the bleeding edge version required by Git for Windows.

To work around that, hard link (or copy, if hard linking fails) the
`.dll`s from `/mingw64/bin` to `/mingw64/libexec/git-core` if they are
conflicting with system ones. For example, if a dll named "ssleay32.dll"
in `/mingw64/bin` is also present in `C:\Windows\System32`, the version
in `/mingw64/bin` is copied to `/mingw64/libexec/git-core`. This call
ensures that the `.dll` in `/mingw64/libexec/git-core` is picked first
when Windows load dll dependencies for executables in
`/mingw64/libexec/git-core`.

This fixes https://github.com/git-for-windows/git/issues/145.

[jes: adjusted commit message, use `{sys}` instead of
`%SYSTEMROOT%\system32` to support 32-bit Git on 64-bit Windows, make
sure that copied `.dll` files are deleted at uninstall.]

Signed-off-by: Matthieu Vachon <matthieu.o.vachon@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>