portable: hard-link or copy all .dll files into exec path
Due to Windows' DLL search order, `.dll` files required by Git's
executables may be picked up from Windows' `%SYSTEMROOT%`:
https://msdn.microsoft.com/en-us/library/windows/desktop/ms682586

This is particularly problematic when an outdated version of, say,
`ssleay32.dll` lives in Windows' system directory, and is picked up when
running `git-remote-https.exe` instead of our own version of said
library.

To work around that, we hard link (or copy, if hard linking fails) *all*
`.dll` files in `/mingw64/bin` to `/mingw64/libexec/git-core/` (or the
equivalent directories in 32-bit setups). This solves the problem
because the directory in which the executable lives itself takes *even
higher* precendence than the Windows system directory.

This patch is a companion to db815ec (installer: Avoid git for windows
DLLs conflicts with system ones, 2015-05-22) for the portable Git. The
only difference is that it hard links *all* `.dll` files, to account for
the fact that a portable Git should be independent of the contents of
the current Windows system directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>