installer: work around misquoting in `Git Bash Here`
When calling `Git Bash Here` from the background of the `C:\` directory
in Windows Explorer, the term `%v` is expanded to the directory. To
allow for spaces in the directory name, we quote that directory when
passing the directory's path as `"--cd=%v"` parameter to `git-bash.exe`.
Unfortunately, the `CommandLineToArgvW()` function mistakes the last
part of `"--cd=C:\"` for a quoted double quote, and as a consequence
`CreateProcess()` fails because there is no `C:"` directory.

To work around this, let's just append `.`: this turns the parameter
into `"--cd=C:\."` which is correctly interpreted as `C:\.` directory.
It works also in the general case, where `%v` does *not* expand into a
path with a trailing backslash because on Windows, `<name>` and
`<name>.` refer to the same file.

This fixes https://github.com/git-for-windows/git/issues/176.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>