installer (64-bit): remove 2GB pack size limit
In 64-bit setups it does not make sense to limit the pack size to 2GB.
So let's remove that setting in such a setup.

Care needs to be taken not to override any different pack size limit
that the user might have chosen before installing Git for Windows.

This fixes https://github.com/git-for-windows/git/issues/288.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>