sdk-installer: support NTLM proxies
When Pacman needs to be told to use a proxy, it actually needs to use a
helper because Pacman's HTTP support has no clues about authentication:
the proxy might require NTLM (or other) authentication.

So let's ship with cURL, and fall back to using it with the `--anyauth`
option.

This fixes https://github.com/git-for-windows/git/issues/309

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>