pacman-mirror: allow 'push' even if only the databases changed
We will introduce the 'remove' command in the next commit. This command
will not require any new version of any package to be uploaded, yet it
still will need the databases to be uploaded.

Prepare the 'push' command to detect that situation.

The diff for this commit is best viewed with the `-w` option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>