installer: try harder to avoid finding erroneous .dll files in the PATH
We used to hard-link or copy all .dll files into libexec/git-core/ when
they were also in the system directory, to avoid problems with the
search path order for DLLs.

However, if users prefix their PATH (e.g. via their `~/.bashrc`), they
might find different DLLs than Git for Windows' yet, even if the DLLs
are actually no in the system32 directory.

So let's just go wild and hard-link (or copy, if hardlinking is not an
option) all the .dll files into our libexec/git-core/ directory so that
the .dll files are definitely in the same directory as the .exe files
that need them.

This fixes https://github.com/git-for-windows/git/issues/318

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>