git-extra: do not let MSys2 read groups from ActiveDirectory
In https://github.com/git-for-windows/git/issues/193 it has been
reported that Git Bash takes an extraordinary time to start up under
certain circumstances. One of those situtations is when the computer is
in a domain but the domain controller cannot be reached.

This developer finally had a chance to reproduce this issue and
debugging revealed that the groups are always read from LDAP, in this
particular case resulting in 30sec delays or longer.

Since Git for Windows does not actually care about groups all that much,
let's just skip reading groups from Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>