portable/installer: work around 32-bit issue with Windows 10
At least on one machine this developer has access to, the default base
address of the 32-bit `msys-2.0.dll` is no good on Windows 10, so we need
to adjust it.

It would be best, of course, if the Batch script could run any program to
guess the best base address, or whether it needs to be adjusted to begin
with, but so far none was identified.

Therefore, let's just stay with the hard-coded base address adjustment
when running in 32-bit mode on Windows 10, until a brilliant idea comes
along to make it better.

This fixes https://github.com/git-for-windows/git/issues/361

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>