git-extra: fix arrow keys in Bash's history in the Windows console
When running Bash in the Windows console (i.e. not in mintty), the reverse
search through the history did not quite work due to the arrow keys not
being mapped correctly.

This fixes https://github.com/git-for-windows/git/issues/353 (probably, at
time of writing, there was no feedback from the reporter yet).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>