git-extra: avoid tilde expansion during tab completion
When connected to an Active Directory (a variant of LDAP), tab completion
becomes very slow because the tilde is expanded all the time, even if
completing paths that do not contain a tilde at all. This developer
observed delays of several seconds at a time.

This is completely unnecessary, so let's do not do that.

The patch was provided by Ove Risberg with a bug report to the original
bash-completion project:

https://bugs.launchpad.net/ubuntu/+source/bash-completion/+bug/1390061

There is already a Pull Request incorporating these changes
(https://github.com/Alexpux/MSYS2-packages/pull/343); Until then, this
commit will work around the issue.

This fixes https://github.com/git-for-windows/git/issues/377

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>