Convert pdf, doc and docx files to text by default
Converting PDF and  Word files to text before diffing them allows an easier comparison between changed
files. This reintroduces some functionality of Git for Windows 1.x.
Including only unzip.exe instead of the entire unzip package makes the installer increase only by 61 kiB
instead of 84 kiB, hence the we opted for the former. pdftotext exists in the xpdf package (adds 2860 kiB) and
the poppler package (adds 13250 kiB), we opted to include the xpdf pdftotext.exe and its dependency
libstdc++-6.dll that add 550 kiB to the installer instead of the poppler pdftotext.exe and its 23 additional dlls,
that would increase the size of the installer by 2032 kiB.
In total this commit increases the size of the installer by 2220 kiB.

This fixes https://github.com/git-for-windows/git/issues/355 .

Signed-off-by: Matthias Aßhauer <mha1993@live.de>