portable: mention post-install.bat in the README
To preempt potential DLL base address as well as DLL search order
problems, the `post-instal.bat` script should really be executed after
unpacking the archive. This is done as part of the installer, but if
somebody decides that they want to unpack the archive themselves, they
should know that this script needs to run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>