git-extra: add aliases to support interactive consoles in mintty
This patch adds support for a couple of interactive consoles that users
invariably want to run in their Git Bash (despite our clear warning in the
release notes and the installer; Who reads text anyways?).

Of course, the list is non-exhaustive. The idea is that we can extend this
whenever needed. Or even better: users can extend this as needed and make
a Pull Request, in true Open Source style facilitated by... *drumroll*...
Git.

This fixes https://github.com/git-for-windows/git/issues/399 and
https://github.com/git-for-windows/git/issues/400

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>