git-extra: build and package WhoUses.exe
Git for Windows 1.x shipped with a pretty useful tool called `WhoUses`
that can identify processes which hold a lock to a given file. This comes
in pretty handy when one would like to remove or rename a file but Windows
refuses to do that because of such a locking process.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>